overlap and chunk_size to be divisible by stride and scale_factor and in fractional pipelines

+++ dorado/basecall/nn/TxModel.h
@@ -50,7 +50,7 @@ struct RotaryEmbeddingImpl : torch::nn::Module {
 
     at::Tensor forward(const at::Tensor &x);
 
-    const int dim, max_seq_len;
+    const int64_t dim, max_seq_len;
     const float theta;
     const at::TensorOptions options;
 };
